+++ src/main/java/blackjack/dto/ProfitResult.java
@@ -2,5 +2,5 @@ package blackjack.dto;
 
 import java.util.List;
 
-public record ProfitResult(double dealerProfit, List<PlayerProfitResult> playerProfits) {
+public record ProfitResult(int dealerProfit, List<PlayerProfitResult> playerProfits) {
 }
